getOrganizationEndpoints moved to test utils
