+++ packages/quantic/force-app/examples/main/lwc/caseAssistUtils/caseAssistUtils.js
@@ -1,7 +1,16 @@
 // @ts-ignore
 import getHeadlessConfiguration from '@salesforce/apex/HeadlessController.getHeadlessConfiguration';
-// @ts-ignore
-import {getOrganizationEndpoints} from 'c/organizationUtils';
+
+export function getOrganizationEndpoints(orgId, env = 'prod') {
+  const envSuffix = env === 'prod' ? '' : env;
+
+  const platform = `https://${orgId}.org${envSuffix}.coveo.com`;
+  const analytics = `https://${orgId}.analytics.org${envSuffix}.coveo.com`;
+  const search = `${platform}/rest/search/v2`;
+  const admin = `https://${orgId}.admin.org${envSuffix}.coveo.com`;
+
+  return {platform, analytics, search, admin};
+}
 
 export async function getCaseAssistId(configName) {
   try {
