+++ packages/quantic/force-app/main/default/lwc/quanticUtils/quanticUtils.js
@@ -34,7 +34,6 @@ import pastYear_plural from '@salesforce/label/c.quantic_PastYear_plural';
 
 export * from './recentQueriesUtils';
 export * from './markdownUtils';
-export * from './quanticOrganizationsUtils';
 
 export class Debouncer {
   _timeout;
